RA-92: Changed output data from Double to Integer to maintain consistency

+++ api/src/main/java/org/cardanofoundation/rosetta/api/construction/service/impl/ConstructionApiServiceImpl.java
@@ -77,7 +77,7 @@ public class ConstructionApiServiceImpl implements ConstructionApiService {
     NetworkIdentifier networkIdentifier = constructionPreprocessRequest.getNetworkIdentifier();
     Optional<ConstructionPreprocessMetadata> metadata = Optional.ofNullable(
         constructionPreprocessRequest.getMetadata());
-    Double relativeTtl;
+    int relativeTtl;
     DepositParameters depositParameters;
     if(metadata.isPresent()) {
       relativeTtl = cardanoService.checkOrReturnDefaultTtl(metadata.get().getRelativeTtl());
@@ -87,10 +87,10 @@ public class ConstructionApiServiceImpl implements ConstructionApiService {
       depositParameters = cardanoService.getDepositParameters();
     }
 
-    Double transactionSize = cardanoService.calculateTxSize(
+    int transactionSize = cardanoService.calculateTxSize(
         NetworkIdentifierType.findByName(networkIdentifier.getNetwork()),
         constructionPreprocessRequest.getOperations(), 0, depositParameters);
-    Map<String, Double> response = Map.of(Constants.RELATIVE_TTL, relativeTtl,
+    Map<String, Integer> response = Map.of(Constants.RELATIVE_TTL, relativeTtl,
         Constants.TRANSACTION_SIZE,
         transactionSize);
     return new ConstructionPreprocessResponse(response, null);
@@ -133,10 +133,10 @@ public class ConstructionApiServiceImpl implements ConstructionApiService {
     int ttl;
     DepositParameters depositParameters;
     if(metadata.isPresent()) {
-      ttl = cardanoService.checkOrReturnDefaultTtl(metadata.get().getTtl()).intValue();
+      ttl = cardanoService.checkOrReturnDefaultTtl(metadata.get().getTtl());
       depositParameters = Optional.ofNullable(metadata.get().getProtocolParameters()).map(protocolParameters -> new DepositParameters(protocolParameters.getKeyDeposit(), protocolParameters.getPoolDeposit())).orElse(cardanoService.getDepositParameters());
     } else {
-      ttl = Constants.DEFAULT_RELATIVE_TTL.intValue();
+      ttl = Constants.DEFAULT_RELATIVE_TTL;
       depositParameters = cardanoService.getDepositParameters();
     }
